Fix inheritance of values which are DataObject
Previous code would in case of inherited value from the parent object
process the inherited value again through mapping.func or mapping.allowed_types
check. This lead to construction of DataObject with first argument being
DataObject instance instead of data taken from yaml.

With this change, if the value is inherited, it's directly stored and no
further processing is done.

Change-Id: Ia08d9d173ee06aa1a7440c2a7208cbd5c7f5d455