DataObject should not crash when optional parent is not provided.
Prior to this change it was required to always specify a parent even if the
data object should not have any parent.

The traceback was also hard to understand where the issue came from as there
was no information about the document which was subject to the traceback so
it seems this was undesired behaviour and the parent should be optional.

Should the parent not be optional and some value (even None) should be set, I'd
prefer adding additional class property which could enforce this.

Change-Id: I922dacde91c2d1ad924830a43161d58f78225311